Refactor component testing: Moved "Component Test" panel to core
