--- code/addons/test/preview.js
@@ -1,1 +1,0 @@
-export * from './dist/preview';
