+++ code/addons/test/src/index.ts
@@ -1,7 +1,5 @@
 import { definePreview } from 'storybook/preview-api';
 
-import * as addonAnnotations from './preview';
-
-export default () => definePreview(addonAnnotations);
+export default () => definePreview({});
 
 export type { TestParameters } from './types';
