+++ code/addons/test/src/manager.tsx
@@ -1,16 +1,12 @@
 import React, { useState } from 'react';
 
-import { AddonPanel } from 'storybook/internal/components';
 import type { StatusValue } from 'storybook/internal/types';
 import { type Addon_TestProviderType, Addon_TypesEnum } from 'storybook/internal/types';
 
 import { a11yStatusStore, componentTestStatusStore, store } from '#manager-store';
-import type { Combo } from 'storybook/manager-api';
-import { Consumer, addons, types } from 'storybook/manager-api';
+import { addons } from 'storybook/manager-api';
 
 import { GlobalErrorContext, GlobalErrorModal } from './components/GlobalErrorModal';
-import { Panel } from './components/Panel';
-import { PanelTitle } from './components/PanelTitle';
 import { TestProviderRender } from './components/TestProviderRender';
 import {
   A11Y_PANEL_ID,
@@ -146,23 +142,4 @@ addons.register(ADDON_ID, (api) => {
       },
     } satisfies Omit<Addon_TestProviderType<Details>, 'id'>);
   }
-
-  const filter = ({ state }: Combo) => {
-    return {
-      storyId: state.storyId,
-    };
-  };
-
-  addons.add(PANEL_ID, {
-    type: types.PANEL,
-    title: () => <PanelTitle />,
-    match: ({ viewMode }) => viewMode === 'story',
-    render: ({ active }) => {
-      return (
-        <AddonPanel active={!!active}>
-          <Consumer filter={filter}>{({ storyId }) => <Panel storyId={storyId} />}</Consumer>
-        </AddonPanel>
-      );
-    },
-  });
 });
