+++ code/core/package.json
@@ -117,6 +117,19 @@
       "types": "./dist/actions/decorator.d.ts",
       "import": "./dist/actions/decorator.js"
     },
+    "./internal/component-testing": {
+      "types": "./dist/component-testing/index.d.ts",
+      "import": "./dist/component-testing/index.js",
+      "require": "./dist/component-testing/index.cjs"
+    },
+    "./internal/component-testing/preview": {
+      "types": "./dist/component-testing/preview.d.ts",
+      "import": "./dist/component-testing/preview.js",
+      "require": "./dist/component-testing/preview.cjs"
+    },
+    "./internal/component-testing/manager": {
+      "import": "./dist/component-testing/manager.js"
+    },
     "./internal/core-events": {
       "types": "./dist/core-events/index.d.ts",
       "import": "./dist/core-events/index.js",
@@ -330,6 +343,12 @@
       "actions/decorator": [
         "./dist/actions/decorator.d.ts"
       ],
+      "internal/component-testing": [
+        "./dist/component-testing/index.d.ts"
+      ],
+      "internal/component-testing/preview": [
+        "./dist/component-testing/preview.d.ts"
+      ],
       "internal/core-events": [
         "./dist/core-events/index.d.ts"
       ],
@@ -465,6 +484,7 @@
     "@babel/preset-react": "^7.26.3",
     "@babel/traverse": "^7.26.9",
     "@babel/types": "^7.26.8",
+    "@devtools-ds/object-inspector": "^1.1.2",
     "@discoveryjs/json-ext": "^0.5.3",
     "@emotion/cache": "^11.14.0",
     "@emotion/is-prop-valid": "^1.3.1",
@@ -572,6 +592,7 @@
     "slash": "^5.0.0",
     "source-map": "^0.7.4",
     "store2": "^2.14.2",
+    "strip-ansi": "^7.1.0",
     "strip-json-comments": "^5.0.1",
     "telejson": "^7.2.0",
     "tiny-invariant": "^1.3.1",
