+++ code/core/scripts/dts.ts
@@ -36,6 +36,9 @@ async function run() {
     'storybook/internal/channels',
     'storybook/internal/client-logger',
     'storybook/internal/common',
+    'storybook/internal/component-testing',
+    'storybook/internal/component-testing/preview',
+    'storybook/internal/component-testing/manager',
     'storybook/internal/components',
     'storybook/internal/core-events',
     'storybook/internal/core-server',
