+++ code/core/scripts/entries.ts
@@ -22,6 +22,10 @@ export const getEntries = (cwd: string) => {
     define('src/actions/manager.tsx', ['browser'], false, ['react'], [], [], true),
     define('src/actions/decorator.ts', ['browser'], true, ['react'], [], [], true),
 
+    define('src/component-testing/index.ts', ['browser', 'node'], true, ['react'], [], []),
+    define('src/component-testing/preview.ts', ['browser', 'node'], true, ['react'], [], []),
+    define('src/component-testing/manager.tsx', ['browser'], false, ['react'], [], []),
+
     define('src/core-events/index.ts', ['browser', 'node'], true),
     define('src/manager-errors.ts', ['browser'], true),
     define('src/preview-errors.ts', ['browser', 'node'], true),
