+++ code/core/src/component-testing/components/Interaction.stories.tsx
