+++ code/core/src/component-testing/components/Interaction.tsx
