+++ code/core/src/component-testing/components/InteractionsPanel.stories.tsx
