+++ code/core/src/component-testing/components/MethodCall.tsx
@@ -3,6 +3,7 @@ import React, { Fragment } from 'react';
 
 import type { Call, CallRef, ElementRef } from 'storybook/internal/instrumenter';
 
+// eslint-disable-next-line import/no-extraneous-dependencies
 import { ObjectInspector } from '@devtools-ds/object-inspector';
 import { useTheme } from 'storybook/theming';
