+++ code/core/src/component-testing/components/Panel.test.ts
