+++ code/core/src/component-testing/components/Panel.tsx
@@ -24,7 +24,7 @@ import {
   ADDON_ID,
   STATUS_TYPE_ID_COMPONENT_TEST,
   STORYBOOK_ADDON_TEST_CHANNEL,
-} from '../constants';
+} from '../../../../addons/test/src/constants';
 import { InteractionsPanel } from './InteractionsPanel';
 
 const INITIAL_CONTROL_STATES = {
