+++ code/core/src/component-testing/components/PanelTitle.tsx
