+++ code/core/src/component-testing/components/StatusBadge.tsx
