+++ code/core/src/component-testing/components/StatusIcon.stories.tsx
