+++ code/core/src/component-testing/components/StatusIcon.tsx
