+++ code/core/src/component-testing/constants.ts
@@ -1,0 +1,5 @@
+export const ADDON_ID = 'storybook/interactions';
+export const PANEL_ID = `${ADDON_ID}/panel`;
+
+export const DOCUMENTATION_LINK = 'writing-tests/test-addon';
+export const DOCUMENTATION_DISCREPANCY_LINK = `${DOCUMENTATION_LINK}#what-happens-when-there-are-different-test-results-in-multiple-environments`;
