+++ code/core/src/component-testing/index.ts
@@ -1,0 +1,2 @@
+// EMPTY FILE
+export default {};
