+++ code/core/src/component-testing/manager.tsx
@@ -1,0 +1,30 @@
+import React from 'react';
+
+import { AddonPanel } from 'storybook/internal/components';
+
+import { type Combo, Consumer, addons, types } from 'storybook/manager-api';
+
+import { Panel } from './components/Panel';
+import { PanelTitle } from './components/PanelTitle';
+import { ADDON_ID, PANEL_ID } from './constants';
+
+addons.register(ADDON_ID, (api) => {
+  const filter = ({ state }: Combo) => {
+    return {
+      storyId: state.storyId,
+    };
+  };
+
+  addons.add(PANEL_ID, {
+    type: types.PANEL,
+    title: () => <PanelTitle />,
+    match: ({ viewMode }) => viewMode === 'story',
+    render: ({ active }) => {
+      return (
+        <AddonPanel active={!!active}>
+          <Consumer filter={filter}>{({ storyId }) => <Panel storyId={storyId} />}</Consumer>
+        </AddonPanel>
+      );
+    },
+  });
+});
