+++ code/core/src/preview-api/core-annotations.ts
@@ -1,3 +1,4 @@
+import componentTestingAnnotations from 'storybook/internal/component-testing/preview';
 import type { Renderer } from 'storybook/internal/types';
 
 import actionAnnotations from 'storybook/actions/preview';
@@ -10,6 +11,8 @@ export function getCoreAnnotations() {
   return [
     // @ts-expect-error CJS fallback
     (actionAnnotations.default ?? actionAnnotations)(),
+    // @ts-expect-error CJS fallback
+    (componentTestingAnnotations.default ?? componentTestingAnnotations)(),
     // @ts-expect-error CJS fallback
     (testAnnotations.default ?? testAnnotations)(),
   ];
