+++ code/core/src/shared/preview/csf4.ts
@@ -1,3 +1,4 @@
+import componentTestingAnnotations from 'storybook/internal/component-testing/preview';
 import type {
   ComponentAnnotations,
   Meta,
@@ -31,6 +32,8 @@ export function __definePreview<TRenderer extends Renderer>(
         composeConfigs([
           // @ts-expect-error CJS fallback
           (actionAnnotations.default ?? actionAnnotations)(),
+          // @ts-expect-error CJS fallback
+          (componentTestingAnnotations.default ?? componentTestingAnnotations)(),
           // @ts-expect-error CJS fallback
           (testAnnotations.default ?? testAnnotations)(),
           ...(addons ?? []),
